Fix segfault in Node server destructor
